Split read gathering methods into a separate module
